Move google analytics script from _document.js to _app.js
